+++ tests/unit/tile/test_tile_component_manager.py
@@ -681,8 +681,6 @@ class TestStaticSimulator:  # pylint: disable=too-many-public-methods
                 TileSimulator.CHANNELISER_TRUNCATION,
                 [[2] * 512],
             ),
-            ("tile_id", TileSimulator.TILE_ID, [123]),
-            ("station_id", TileSimulator.STATION_ID, [321]),
             ("test_generator_active", False, [True]),
             ("csp_spead_format", TileSimulator.CSP_SPEAD_FORMAT, ["AAVS"]),
         ),
@@ -1116,13 +1114,13 @@ class TestStaticSimulator:  # pylint: disable=too-many-public-methods
 
         # Set tile_id case
         tile_component_manager._station_id = 2
-        tile_component_manager.tile_id = 5
+        tile_component_manager.set_tile_id(5)
         assert tile_simulator._station_id == 2
         assert tile_simulator._tile_id == 5
 
         # Set station_id case
         tile_component_manager._tile_id = 2
-        tile_component_manager.station_id = 5
+        tile_component_manager.set_station_id(5)
         assert tile_simulator._station_id == 5
         assert tile_simulator._tile_id == 2
 
@@ -1134,13 +1132,15 @@ class TestStaticSimulator:  # pylint: disable=too-many-public-methods
         )
         # set station_id with mocked failure
         tile_component_manager._tile_id = initial_tile_id + 1
-        tile_component_manager.station_id = initial_station_id + 1
+        with pytest.raises(LibraryError):
+            tile_component_manager.set_station_id(initial_station_id + 1)
         assert tile_simulator._station_id == initial_station_id
         assert tile_simulator._tile_id == initial_tile_id
 
         # set tile_id with mocked failure
         tile_component_manager._station_id = initial_station_id + 1
-        tile_component_manager.tile_id = initial_tile_id + 1
+        with pytest.raises(LibraryError):
+            tile_component_manager.set_tile_id(initial_tile_id + 1)
         assert tile_simulator._station_id == initial_station_id
         assert tile_simulator._tile_id == initial_tile_id
 
@@ -1965,10 +1965,9 @@ class TestStaticSimulator:  # pylint: disable=too-many-public-methods
         # Set preADU levels to 3 for all channels
         tile_component_manager.set_preadu_levels([4.0] * 32)
         assert tile_simulator.tpm.preadu[1].get_attenuation()[1] == 4.00
-        # Try to set more levels (33) than there are channels (32),
-        # in order to check that the TileComponentManager swallows exceptions.
-        # Possibly a bad idea?
-        tile_component_manager.set_preadu_levels([3.0] * 33)
+
+        with pytest.raises(ValueError):
+            tile_component_manager.set_preadu_levels([3.0] * 33)
 
     def test_load_calibration_coefficients(
         self: TestStaticSimulator,
@@ -2950,8 +2949,6 @@ class TestStaticSimulator:  # pylint: disable=too-many-public-methods
         ("attribute"),
         [
             ("register_list"),
-            ("station_id"),
-            ("tile_id"),
             ("is_programmed"),
             ("firmware_version"),
             ("firmware_name"),